+++ examples/with-gin/config/config.go
@@ -1,6 +1,7 @@
 package config
 
 import (
+	"github.com/supertokens/supertokens-golang/recipe/dashboard"
 	"log"
 
 	"github.com/spf13/viper"
@@ -105,6 +106,7 @@ func Init() {
 				},
 			}),
 			session.Init(nil),
+			dashboard.Init(nil),
 			// thirdparty.Init(thirdpartyConfig),
 		},
 	})