+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 -   Adds unit test for Apple callback form post
+-   Updates all example apps to also initialise dashboard recipe
 
 ## [0.10.2] - 2023-02-24
 -   Adds APIs and logic to the dashboard recipe to enable email password based login